move unit tests (doesn't depend on SwingSet) to unitTest folder
